[Rename] : components PascalCase로 네이밍 변경
- SpeedGame, StartGame, GameInputs 변경
