+++ components/modules/GameInputs.tsx
